requirements.txt
torch==1.12.1
torchaudio==0.13.1
numpy==1.24.1
scipy==1.10.0
PyYAML==5.3.1
tensorboard==2.11.2
tqdm==4.64.1
scikit-learn==1.2.0
